refactor: :recycle: convert plugin-browser-device to typescript es module
