+++ packages/plugin-browser-device/rollup.config.npm.mjs
@@ -1,0 +1,6 @@
+import createRollupConfig from '../../.rollup/index.mjs'
+
+export default createRollupConfig({
+    input: 'src/device.ts',
+    external: ['@bugsnag/cuid'],
+})
