+++ packages/plugin-browser-device/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../tsconfig.json",
+    "include": ["src/**/*.ts"],
+    "compilerOptions": {
+      "target": "ES2020"
+    }
+}
+
